# src/backend/openapi/api-contracts-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EasyPassword API",
    "description": "Passwordless vault with WebAuthn and end-to-end encryption",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/vault/": {
      "get": {
        "tags": [
          "vault"
        ],
        "summary": "List Vault",
        "operationId": "list_vault_api_v1_vault__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VaultItemResponse"
                  },
                  "type": "array",
                  "title": "Response List Vault Api V1 Vault  Get"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed or reauthentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "ReauthenticationRequired",
                  "message": "Reauthentication required"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "vault"
        ],
        "summary": "Create Vault",
        "operationId": "create_vault_api_v1_vault__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VaultCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultItemResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed or reauthentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "ReauthenticationRequired",
                  "message": "Reauthentication required"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error - request violated the Phase 3.1 contract (e.g., extra fields, suspicious key names, invalid blob, blob too large, unsupported blob version).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "extraFields": {
                    "summary": "Extra fields rejected",
                    "value": {
                      "code": "ERR_EXTRA_FIELDS",
                      "message": "Unknown fields are not allowed",
                      "details": {
                        "fields": ["plaintext_password"]
                      }
                    }
                  },
                  "suspiciousKey": {
                    "summary": "Plaintext-looking field rejected",
                    "value": {
                      "code": "ERR_SUSPICIOUS_KEY",
                      "message": "Field name suggests plaintext credential; client-side encryption required",
                      "details": {
                        "field": "password"
                      }
                    }
                  },
                  "invalidBlob": {
                    "summary": "Invalid blob format",
                    "value": {
                      "code": "ERR_INVALID_BLOB",
                      "message": "Blob is not valid base64 or does not match the expected layout"
                    }
                  },
                  "blobTooLarge": {
                    "summary": "Blob exceeds size limit",
                    "value": {
                      "code": "ERR_BLOB_TOO_LARGE",
                      "message": "Blob exceeds the maximum allowed size of 1048576 bytes"
                    }
                  },
                  "unsupportedVersion": {
                    "summary": "Unsupported blob version",
                    "value": {
                      "code": "ERR_UNSUPPORTED_BLOB_VERSION",
                      "message": "Blob version 2 is not supported by this server"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vault/{vault_id}": {
      "put": {
        "tags": [
          "vault"
        ],
        "summary": "Update Vault",
        "operationId": "update_vault_api_v1_vault__vault_id__put",
        "parameters": [
          {
            "name": "vault_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vault Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VaultUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultItemResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed or reauthentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "ReauthenticationRequired",
                  "message": "Reauthentication required"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error - request violated the Phase 3.1 contract (e.g., extra fields, suspicious key names, invalid blob, blob too large, unsupported blob version).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "extraFields": {
                    "summary": "Extra fields rejected",
                    "value": {
                      "code": "ERR_EXTRA_FIELDS",
                      "message": "Unknown fields are not allowed",
                      "details": {
                        "fields": ["plaintext_password"]
                      }
                    }
                  },
                  "suspiciousKey": {
                    "summary": "Plaintext-looking field rejected",
                    "value": {
                      "code": "ERR_SUSPICIOUS_KEY",
                      "message": "Field name suggests plaintext credential; client-side encryption required",
                      "details": {
                        "field": "password"
                      }
                    }
                  },
                  "invalidBlob": {
                    "summary": "Invalid blob format",
                    "value": {
                      "code": "ERR_INVALID_BLOB",
                      "message": "Blob is not valid base64 or does not match the expected layout"
                    }
                  },
                  "blobTooLarge": {
                    "summary": "Blob exceeds size limit",
                    "value": {
                      "code": "ERR_BLOB_TOO_LARGE",
                      "message": "Blob exceeds the maximum allowed size of 1048576 bytes"
                    }
                  },
                  "unsupportedVersion": {
                    "summary": "Unsupported blob version",
                    "value": {
                      "code": "ERR_UNSUPPORTED_BLOB_VERSION",
                      "message": "Blob version 2 is not supported by this server"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "vault"
        ],
        "summary": "Delete Vault",
        "operationId": "delete_vault_api_v1_vault__vault_id__delete",
        "parameters": [
          {
            "name": "vault_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vault Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Vault Api V1 Vault  Vault Id  Delete"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed or reauthentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "ReauthenticationRequired",
                  "message": "Reauthentication required"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "VaultCreateRequest": {
        "properties": {
          "service_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Service Name"
          },
          "login_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Login Name"
          },
          "password_blob": {
            "$ref": "#/components/schemas/BlobV1",
            "description": "Client-encrypted password blob (base64-encoded [version][iv][ciphertext+tag])."
          },
          "notes_blob": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BlobV1"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional client-encrypted notes blob (base64-encoded [version][iv][ciphertext+tag])."
          },
          "key_version": {
            "type": "integer",
            "minimum": 1,
            "default": 1,
            "title": "Key Version",
            "description": "Version of the client-side key used to encrypt the blobs. Used for future key rotation."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "service_name",
          "login_name",
          "password_blob",
          "key_version"
        ],
        "title": "VaultCreateRequest",
        "description": "Request schema for creating a vault item. All blobs MUST be client-side encrypted before submission (Phase 3.1 contract). For AES-GCM, the cryptographic contract uses Additional Authenticated Data (AAD) built as UTF8(NFC(user_id)) + 0x1F + UTF8(NFC(item_id)) + 0x1F + UTF8(NFC(service_name)) in the exact field order user_id, item_id, service_name. AAD is part of the encryption contract and is not exposed as a direct API field."
      },
      "VaultItemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "service_name": {
            "type": "string",
            "title": "Service Name"
          },
          "login_name": {
            "type": "string",
            "title": "Login Name"
          },
          "password_blob": {
            "$ref": "#/components/schemas/BlobV1",
            "description": "Client-encrypted password blob (base64-encoded [version][iv][ciphertext+tag])."
          },
          "notes_blob": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BlobV1"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional client-encrypted notes blob (base64-encoded [version][iv][ciphertext+tag])."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "blob_version_detected": {
            "type": "integer",
            "minimum": 1,
            "default": 1,
            "title": "Blob Version Detected",
            "description": "Version of the blob format detected by the server. Used to flag items that may need migration."
          },
          "migration_recommended": {
            "type": "boolean",
            "default": false,
            "title": "Migration Recommended",
            "description": "True when the server detects the blob uses a deprecated version and the client should re-encrypt with the current version."
          },
          "key_version": {
            "type": "integer",
            "minimum": 1,
            "default": 1,
            "title": "Key Version",
            "description": "Version of the client-side key used to encrypt the blobs."
          }
        },
        "type": "object",
        "required": [
          "id",
          "service_name",
          "login_name",
          "password_blob",
          "created_at",
          "updated_at",
          "blob_version_detected",
          "migration_recommended",
          "key_version"
        ],
        "title": "VaultItemResponse",
        "description": "Response schema for a vault item. Blobs are returned as opaque ciphertext; the server never decrypts them. AES-GCM encryption uses Additional Authenticated Data (AAD) built as UTF8(NFC(user_id)) + 0x1F + UTF8(NFC(item_id)) + 0x1F + UTF8(NFC(service_name)) in the exact field order user_id, item_id, service_name. AAD is documented as part of the cryptographic contract and is not returned as a separate API field."
      },
      "VaultUpdateRequest": {
        "properties": {
          "service_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Name"
          },
          "login_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Login Name"
          },
          "password_blob": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BlobV1"
              },
              {
                "type": "null"
              }
            ],
            "description": "Client-encrypted password blob (base64-encoded [version][iv][ciphertext+tag])."
          },
          "notes_blob": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BlobV1"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional client-encrypted notes blob (base64-encoded [version][iv][ciphertext+tag])."
          },
          "key_version": {
            "type": "integer",
            "minimum": 1,
            "default": 1,
            "title": "Key Version",
            "description": "Version of the client-side key used to encrypt the blobs."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "VaultUpdateRequest",
        "description": "Request schema for updating a vault item. All blobs MUST be client-side encrypted before submission (Phase 3.1 contract). For AES-GCM, the cryptographic contract uses Additional Authenticated Data (AAD) built as UTF8(NFC(user_id)) + 0x1F + UTF8(NFC(item_id)) + 0x1F + UTF8(NFC(service_name)) in the exact field order user_id, item_id, service_name. AAD is part of the encryption contract and is not exposed as a direct API field."
      },
      "BlobV1": {
        "type": "string",
        "format": "byte",
        "minLength": 1,
        "maxLength": 1048576,
        "pattern": "^[A-Za-z0-9+/]+={0,2}$",
        "title": "BlobV1",
        "description": "Opaque client-encrypted blob (base64). Layout: [1-byte version][12-byte IV][ciphertext+16-byte GCM tag]. AES-GCM uses Additional Authenticated Data (AAD) built as UTF8(NFC(user_id)) + 0x1F + UTF8(NFC(item_id)) + 0x1F + UTF8(NFC(service_name)) in the exact field order user_id, item_id, service_name. Example: AAD = UTF8(NFC('user')) + 0x1F + UTF8(NFC('item')) + 0x1F + UTF8(NFC('service')). The server treats this as opaque bytes and never decrypts it."
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Canonical error code from the Phase 3.1 contract error catalog."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable error message. Safe to display to end users."
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Optional structured details (e.g., offending field names, expected vs actual values)."
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "description": "Canonical error envelope used by all vault endpoints. See Phase 3.1 contract for the full error catalog."
      },
...
    }
  }
}